+++ exports/metadata.mobile.json/index.js
@@ -1,0 +1,1 @@
+export * from '../../metadata.mobile.json.js'