+++ max/metadata.js
@@ -1,7 +1,7 @@
 // Importing from `.json.js` a workaround for a bug in web browsers' "native"
 // ES6 importing system which is uncapable of importing "*.json" files.
 // https://github.com/catamphetamine/libphonenumber-js/issues/239
-import metadata from '../metadata.full.json.js'
+import metadata from '../metadata.max.json.js'
 export default metadata
 
 export function withMetadata(func, _arguments) {