+++ test/metadata/1.1.11/test/types.test.js
@@ -1,4 +1,4 @@
-import metadata from '../metadata.full'
+import metadata from '../metadata.max.json'
 import Metadata from '../../../../source/metadata'
 import get_number_type_custom from '../../../../source/getNumberType'
 import { checkNumberLengthForType } from '../../../../source/helpers/checkNumberLength'