+++ test/metadata/1.7.34/test/getNumberType.test.js
@@ -1,4 +1,4 @@
-import metadata from '../metadata.full'
+import metadata from '../metadata.max.json'
 import Metadata from '../../../../source/metadata'
 import { checkNumberLengthForType } from '../../../../source/helpers/checkNumberLength'
 import _getNumberType from '../../../../source/getNumberType'